requirements.txt
python-dotenv~=0.14.0
requests~=2.24.0
selenium~=3.141.0